requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
ipython==7.31.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
openapi-codec==1.3.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
